% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApollonianGasket-data.R
\docType{data}
\name{ApollonianGasket}
\alias{ApollonianGasket}
\title{Apollonian gasket fractal.}
\format{An object of class \code{'list'} of length 3.}
\usage{
ApollonianGasket
}
\description{
List of contractions for Apollonian gasket.
}
\examples{
plot(createIFS(ApollonianGasket), 7)
}
\keyword{datasets}
