% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPoints.R
\name{colorPoints}
\alias{colorPoints}
\title{Coloring fractal points.}
\usage{
colorPoints(iter, lenContr, lenStartPts = 1)
}
\arguments{
\item{iter}{Number of iterations that should be performed in the \code{plot.IFS()} function.}

\item{lenContr}{Number of contractions used in given IFS.}

\item{lenStartPts}{Length of starting points list.}
}
\value{
List of colors that corresponds to contractions.
}
\description{
Creating list of colors for the \code{plot.IFS()} function such that one color corresponds to one contraction.
}
