% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIFS.R
\name{createIFS}
\alias{createIFS}
\title{Creating object of the IFS class.}
\usage{
createIFS(contractions)
}
\arguments{
\item{contractions}{List of contractions for fractal generation.}
}
\value{
Object of IFS class.
}
\description{
The IFS class is dedicated to create and plot fractals
that can be generated from systems of contractions (IFS - Iterated Functions System).
}
\examples{
myFractal <- createIFS(SierpinskiTriangle)
}
