% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPoints.R
\name{createPoints}
\alias{createPoints}
\title{Creating the list of points to corresponding given fractal.}
\usage{
createPoints(contractions, iter, startPts = list(c(0, 0)))
}
\arguments{
\item{contractions}{List of contractions that generate given fractal.}

\item{iter}{Number of iterations that should be executed.}

\item{startPts}{List of starting points.}
}
\value{
List of points that shapes given fractal.
}
\description{
Each points is represented via its standard Cartesian coordinates.
}
