% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFSpremium.R
\docType{package}
\name{IFSpremium-package}
\alias{IFSpremium}
\alias{IFSpremium-package}
\title{IFSpremium}
\description{
The package IFSpremium allows creating objects of IFS class and plotting them.
IFS fractals are fractals generated by systems of contractions (i.e. functions. with some specific properties).
A list of most popular fractals (e.g. Sierpinski Triangle, Koch Curve) is included in the package so one can easily
plot a fractal via plot() for objects of IFS class and enjoy its beauty.
}
\details{
Included fractals in dataset:
\itemize{
 \item{Apollonian Gasket}
 \item{Koch Curve}
 \item{Koch Snowflake}
 \item{Sierpinski Triangle}
 \item{Sierpinski Carpet}
 \item{Sierpinski Pentagon}
 \item{Pythagorean Tree}
}
}
\author{
\strong{Maintainer}: Michał Futrega \email{michal.futrega@gmail.com}

Authors:
\itemize{
  \item Zofia Grochulska \email{zosia.grochulska@gmail.com}
}

}
