% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KochCurve-data.R
\docType{data}
\name{KochCurve}
\alias{KochCurve}
\title{Koch curve fractal.}
\format{An object of class \code{'list'} of length 4.}
\usage{
KochCurve
}
\description{
List of contractions for Koch curve.
}
\examples{
plot(createIFS(KochCurve), 7)
}
\keyword{datasets}
