% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KochSnowflake-data.R
\docType{data}
\name{KochSnowflake}
\alias{KochSnowflake}
\title{Koch snowflake fractal.}
\format{An object of class \code{'list'} of length 7.}
\usage{
KochSnowflake
}
\description{
List of contractions for Koch snowflake.
}
\examples{
plot(createIFS(KochSnowflake), 6)
}
\keyword{datasets}
