% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PythagoreanTree-data.R
\docType{data}
\name{PythagoreanTree}
\alias{PythagoreanTree}
\title{Pythagorean tree fractal.}
\format{An object of class \code{'list'} of length 3.}
\usage{
PythagoreanTree
}
\description{
List of contractions for Pythagorean tree.
Note: to get desired result use square as starting points.
}
\examples{
points <- seq(0, 1, 1/10)
square <- list()
for (i in seq_along(1:11))
  for (j in seq_along(1:11))
    square <- append(square, list(c(points[i], points[j])))
plot(createIFS(PythagoreanTree), 5, square)
}
\keyword{datasets}
