% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SierpinskiCarpet-data.R
\docType{data}
\name{SierpinskiCarpet}
\alias{SierpinskiCarpet}
\title{Sierpinski carpet fractal.}
\format{An object of class \code{'list'} of length 8.}
\usage{
SierpinskiCarpet
}
\description{
List of contractions for Sierpinski carpet.
}
\examples{
plot(createIFS(SierpinskiCarpet), 5)
}
\keyword{datasets}
