% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SierpinskiPentagon-data.R
\docType{data}
\name{SierpinskiPentagon}
\alias{SierpinskiPentagon}
\title{Sierpinski pentagon fractal.}
\format{An object of class \code{'list'} of length 5.}
\usage{
SierpinskiPentagon
}
\description{
List of contractions for Sierpinski pentagon.
}
\examples{
plot(createIFS(SierpinskiPentagon), 6)
}
\keyword{datasets}
