% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SierpinskiTriangle-data.R
\docType{data}
\name{SierpinskiTriangle}
\alias{SierpinskiTriangle}
\title{Sierpinski triangle fractal.}
\format{An object of class \code{'list'} of length 3.}
\usage{
SierpinskiTriangle
}
\description{
List of contractions for Sierpinski triangle.
}
\examples{
plot(createIFS(SierpinskiTriangle), 8)
}
\keyword{datasets}
